// src/app/models/corporate.ts
export interface Corporate {
      CorporateCustId: number;
      CorporateNumber: string;
      CompanyName: string;
      NatureOfBusiness: string;
      RegistrationNumber: string;
      DateOfIncorporation: Date;
      CompanyLocation: string;
      CompanyPhone: string;
      CompanyFax: string;
      CompanyEmail: string;
      CompanyWebsite: string;
      TinNumber: string;
      RelationalOfficer: number;
      Placement: boolean;
      Address: string;
      BroadcastAlert: boolean;
      SupportAlert: boolean;
}